// TextNotesClient/text-notes-client/out/models/AllGet200ResponseInner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenApi manifest for my TextNoteClient
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AllGet200ResponseInner
 */
export interface AllGet200ResponseInner {
    /**
     * 
     * @type {string}
     * @memberof AllGet200ResponseInner
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof AllGet200ResponseInner
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof AllGet200ResponseInner
     */
    date?: string;
    /**
     * 
     * @type {string}
     * @memberof AllGet200ResponseInner
     */
    content?: string;
}

/**
 * Check if a given object implements the AllGet200ResponseInner interface.
 */
export function instanceOfAllGet200ResponseInner(value: object): value is AllGet200ResponseInner {
    if (!('id' in value) || value['id'] === undefined) return false;
    return true;
}

export function AllGet200ResponseInnerFromJSON(json: any): AllGet200ResponseInner {
    return AllGet200ResponseInnerFromJSONTyped(json, false);
}

export function AllGet200ResponseInnerFromJSONTyped(json: any, ignoreDiscriminator: boolean): AllGet200ResponseInner {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'title': json['title'] == null ? undefined : json['title'],
        'date': json['date'] == null ? undefined : json['date'],
        'content': json['content'] == null ? undefined : json['content'],
    };
}

export function AllGet200ResponseInnerToJSON(json: any): AllGet200ResponseInner {
    return AllGet200ResponseInnerToJSONTyped(json, false);
}

export function AllGet200ResponseInnerToJSONTyped(value?: AllGet200ResponseInner | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'date': value['date'],
        'content': value['content'],
    };
}